Add agent API with server-side ClickHouse tools
