+++ src/components/chat/chat-factory.ts
@@ -4,10 +4,14 @@ import { ModelManager } from "@/components/settings/models/model-manager";
 import type { PlanToolOutput } from "@/lib/ai/agent/plan/planning-types";
 import type { AgentContext, AppUIMessage, Message, MessageMetadata } from "@/lib/ai/ai-types";
 import { sanitizeMessageForPersistence } from "@/lib/ai/session/serialization";
-import { ClientToolExecutors } from "@/lib/ai/tools/client/client-tool-executors";
-import type { StageStatus, ToolProgressCallback } from "@/lib/ai/tools/client/client-tool-types";
+import { ClickHouseToolExecutors } from "@/lib/ai/tools/clickhouse/clickhouse-tool-executors";
+import type {
+  StageStatus,
+  ToolProgressCallback,
+} from "@/lib/ai/tools/clickhouse/clickhouse-tool-types";
+import type { ClickHouseConnection } from "@/lib/ai/tools/clickhouse/clickhouse-tools";
+import { useToolProgressStore } from "@/lib/ai/tools/clickhouse/tool-progress-store";
 import { CLIENT_TOOL_NAMES } from "@/lib/ai/tools/client/client-tools";
-import { useToolProgressStore } from "@/lib/ai/tools/client/tool-progress-store";
 import { SERVER_TOOL_NAMES } from "@/lib/ai/tools/server/server-tool-names";
 import { BasePath } from "@/lib/base-path";
 import { Connection, type QueryResponse } from "@/lib/connection/connection";
@@ -26,7 +30,7 @@ type AbortableQueryResult<TResponse extends QueryResponse | Response> = {
   response: Promise<TResponse>;
   abortController: AbortController;
 };
-type ClientToolName = keyof typeof ClientToolExecutors;
+type ClickHouseExecutorName = keyof typeof ClickHouseToolExecutors;
 const PROVISIONAL_SESSION_TITLE_WORDS = 8;
 
 type ChatFactoryCreateOptions = {
@@ -71,6 +75,7 @@ type SendMessagesRequestPayloadArgs = {
   messageId: string | undefined;
   body: unknown;
   requestContext?: DatabaseContext;
+  clickHouseConnection?: ClickHouseConnection;
   currentModel?: {
     provider: string;
     modelId: string;
@@ -145,6 +150,7 @@ export function buildSendMessagesRequestPayload({
   messageId,
   body,
   requestContext,
+  clickHouseConnection,
   currentModel,
   generateTitle,
   ephemeral,
@@ -169,6 +175,7 @@ export function buildSendMessagesRequestPayload({
         pruneValidateSql,
         outputReasoning,
       },
+      ...(clickHouseConnection ? { connection: clickHouseConnection } : {}),
       ...(requestContext ? { context: requestContext } : {}),
       ...(currentModel ? { model: currentModel } : {}),
     };
@@ -184,9 +191,24 @@ export function buildSendMessagesRequestPayload({
       pruneValidateSql,
       outputReasoning,
     },
+    ...(clickHouseConnection ? { connection: clickHouseConnection } : {}),
     generateTitle,
     ...(requestContext ? { context: requestContext } : {}),
     ...(currentModel ? { model: currentModel } : {}),
+  };
+}
+
+function buildClickHouseConnectionPayload(
+  connection: Connection | null
+): ClickHouseConnection | undefined {
+  if (!connection || !connection.password) {
+    return undefined;
+  }
+
+  return {
+    url: connection.url,
+    user: connection.user,
+    password: connection.password,
   };
 }
 
@@ -451,7 +473,7 @@ export class ChatFactory {
       transport: new DefaultChatTransport({
         fetch: async (_input, init) => {
           const mode = AgentConfigurationManager.getConfiguration().mode;
-          const endpoint = BasePath.getURL(mode === "v2" ? "/api/ai/chat/v2" : "/api/ai/chat");
+          const endpoint = BasePath.getURL(mode === "v2" ? "/api/ai/agent" : "/api/ai/chat");
           return fetch(endpoint, init);
         },
 
@@ -476,6 +498,10 @@ export class ChatFactory {
 
           const requestContext = options.context ?? ChatContext.build();
           const chatPersistenceMode = getRuntimeConfig().sessionRepositoryType;
+          const clickHouseConnection =
+            AgentConfigurationManager.getConfiguration().mode === "v2"
+              ? buildClickHouseConnectionPayload(connection)
+              : undefined;
           return {
             body: buildSendMessagesRequestPayload({
               sessionId,
@@ -485,6 +511,7 @@ export class ChatFactory {
               messageId,
               body,
               requestContext,
+              clickHouseConnection,
               currentModel,
               generateTitle: options.generateTitle,
               ephemeral: options.ephemeral,
@@ -522,7 +549,7 @@ export class ChatFactory {
           return;
         }
 
-        if (!(toolName in ClientToolExecutors)) {
+        if (!(toolName in ClickHouseToolExecutors)) {
           console.error(`Unknown tool: ${toolName}`);
           chat.addToolOutput({
             tool: toolName as never,
@@ -532,7 +559,7 @@ export class ChatFactory {
           return;
         }
 
-        const executor = ClientToolExecutors[toolName as ClientToolName];
+        const executor = ClickHouseToolExecutors[toolName as ClickHouseExecutorName];
 
         try {
           if (!clientToolConnection) {
