+++ src/components/chat/message/message-tool-get-tables.tsx
@@ -7,7 +7,7 @@ import {
   TableRow,
 } from "@/components/ui/table";
 import type { AppUIMessage, ToolPart } from "@/lib/ai/ai-types";
-import type { GetTablesInput, GetTablesOutput } from "@/lib/ai/tools/client/get-tables";
+import type { GetTablesInput, GetTablesOutput } from "@/lib/ai/tools/clickhouse/get-tables";
 import { memo } from "react";
 import { CollapsiblePart } from "./collapsible-part";
