+++ src/lib/ai/agent/general-agent.ts
@@ -1,6 +1,6 @@
 import { streamText, type ModelMessage } from "ai";
 import { LanguageModelProviderFactory } from "../llm/llm-provider-factory";
-import { ClientTools as clientTools } from "../tools/client/client-tools";
+import { ClickHouseTools as clickHouseTools } from "../tools/clickhouse/clickhouse-tools";
 import { SERVER_TOOL_NAMES } from "../tools/server/server-tool-names";
 import type { ServerDatabaseContext } from "./common-types";
 import type { InputModel } from "./plan/sub-agent-registry";
@@ -101,12 +101,12 @@ Guidelines:
       ...messages,
     ],
     tools: {
-      get_tables: clientTools.get_tables,
-      explore_schema: clientTools.explore_schema,
+      get_tables: clickHouseTools.get_tables,
+      explore_schema: clickHouseTools.explore_schema,
       [SERVER_TOOL_NAMES.GENERATE_SQL]: createGenerateSqlTool(modelConfig, context),
-      validate_sql: clientTools.validate_sql,
-      execute_sql: clientTools.execute_sql,
-      search_query_log: clientTools.search_query_log,
+      validate_sql: clickHouseTools.validate_sql,
+      execute_sql: clickHouseTools.execute_sql,
+      search_query_log: clickHouseTools.search_query_log,
     },
   });
 }
