+++ src/lib/ai/ai-types.ts
@@ -1,4 +1,5 @@
 import type { PlannerMetadata } from "@/lib/ai/agent/plan/planning-types";
+import type { ClickHouseTools } from "@/lib/ai/tools/clickhouse/clickhouse-tools";
 import type { ClientTools } from "@/lib/ai/tools/client/client-tools";
 import type { InferUITools, LanguageModelUsage, UIDataTypes, UIMessage } from "ai";
 
@@ -136,11 +137,12 @@ export interface Chat {
  * App UI message: UIMessage with MessageMetadata plus UI timestamps.
  * Single source of truth for message metadata (usage, planner) shared with Message.
  */
-export type AppUIMessage = UIMessage<
-  MessageMetadata,
-  UIDataTypes,
-  InferUITools<typeof ClientTools>
-> & { updatedAt?: Date; createdAt?: Date };
+type AppTools = typeof ClientTools & typeof ClickHouseTools;
+
+export type AppUIMessage = UIMessage<MessageMetadata, UIDataTypes, InferUITools<AppTools>> & {
+  updatedAt?: Date;
+  createdAt?: Date;
+};
 
 /**
  * Type for tool parts that have input, output, and state properties.
