+++ src/lib/ai/session/remote-chat-request.ts
@@ -1,11 +1,13 @@
 import type { ServerDatabaseContext } from "@/lib/ai/agent/common-types";
 import type { AgentContext, AppUIMessage } from "@/lib/ai/ai-types";
+import type { ClickHouseConnection } from "@/lib/ai/tools/clickhouse/clickhouse-tools";
 
 export interface ChatRequestBase {
   sessionId: string;
   connectionId: string;
   ephemeral?: boolean;
   context?: ServerDatabaseContext;
+  connection?: ClickHouseConnection;
   model?: { provider: string; modelId: string; apiKey?: string };
   agentContext?: AgentContext;
 }
