+++ src/lib/ai/tools/clickhouse/clickhouse-tool-types.ts
