+++ src/lib/ai/tools/clickhouse/collect-sql-optimization-evidence.test.ts
