+++ src/lib/ai/tools/clickhouse/find-expensive-queries.ts
@@ -6,7 +6,7 @@
  */
 import { QueryError, type JSONCompactFormatResponse } from "@/lib/connection/connection";
 import { SqlUtils } from "@/lib/sql-utils";
-import type { ToolExecutor } from "./client-tool-types";
+import type { ToolExecutor } from "./clickhouse-tool-types";
 
 const METRIC_CONFIG = {
   cpu: {
