+++ src/lib/ai/tools/clickhouse/rca/evidence-collector-extension.ts
