+++ src/lib/ai/tools/clickhouse/rca/evidence-collector-factory.ts
