+++ src/lib/ai/tools/clickhouse/rca/impl/template-based-collector.ts
