+++ src/lib/ai/tools/clickhouse/rca/tool-collect-rca-evidence.test.ts
