+++ src/lib/ai/tools/clickhouse/search-query-log.test.ts
