+++ src/lib/ai/tools/clickhouse/status/collect-cluster-status.ts
@@ -1,5 +1,5 @@
 import { QueryError, type JSONCompactFormatResponse } from "@/lib/connection/connection";
-import type { ToolExecutor, ToolProgressCallback } from "../client-tool-types";
+import type { ToolExecutor, ToolProgressCallback } from "../clickhouse-tool-types";
 import {
   getSystemMetrics,
   type GetSystemMetricsOutput,
